refactor : 리팩토링 내용 적용 (#100)
