--- src/components/Element/User/style/User.scss
@@ -1,18 +1,0 @@
-@import '/src/styles/scss/mixin';
-
-/* 게시물 */
-.user-container {
-  display: flex;
-  align-items: center;
-  width: fit-content;
-
-  .profile-img {
-    margin-right: 0.75rem;
-    box-shadow: 0 0 0 0.0625rem inset #dbdbdb;
-  }
-
-  .user-info {
-    margin-right: 0.75rem;
-    // display: inline-block;
-  }
-}
