--- src/components/Element/User/style/UserInfo.scss
@@ -1,49 +1,0 @@
-@import '/src/styles/scss/mixin';
-
-.user-info {
-  h3 {
-    margin-bottom: 6px;
-  }
-  .user-name {
-    font-size: map-get($font-sizes, 'size14');
-    font-weight: 500;
-    line-height: 1.125rem;
-    @include get-color(color, font-color);
-  }
-  .user-id {
-    @include get-color(color, gray-color);
-    font-size: map-get($font-sizes, 'size12');
-    font-weight: 400;
-
-    margin-top: 0.125rem;
-    &:before {
-      content: '@ ';
-    }
-  }
-
-  .user-intro {
-    @include get-color(color, gray-color);
-    font-size: map-get($font-sizes, 'size12');
-    font-weight: 400;
-
-    line-height: 0.9375rem;
-    margin-top: 0.375rem;
-    cursor: default;
-    max-width: calc(100vw - 50px - 56px - 32px - 24px);
-  }
-
-  .chat-content {
-    @include get-color(color, gray-color);
-    font-size: map-get($font-sizes, 'size12');
-    font-weight: 400;
-
-    display: block;
-    margin-top: 0.25rem;
-  }
-}
-
-.SubInfoFont {
-  @include get-color(color, gray-color);
-  font-size: map-get($font-sizes, 'size12');
-  font-weight: 400;
-}
